changes required by review

+++ docker/beam-environment/entrypoint.sh
@@ -2,51 +2,56 @@
 
 echo "Starting entrypoint script, at $(date "+%Y-%m-%d-%H:%M:%S")"
 
+function print_error(){
+  echo ""
+  echo "ERROR!"
+  echo "ERROR!"
+  echo "ERROR!"
+  echo "$1"
+  echo ""
+}
 
 ##
 ## print variables that might be set only outside of image
 ##
-required_variables_from_outside=(
-  BEAM_BRANCH_NAME BEAM_COMMIT_SHA BEAM_DATA_BRANCH_NAME BEAM_DATA_COMMIT_SHA
-  BEAM_CONFIG TITLED MAX_RAM
-  INSTANCE_ID INSTANCE_TYPE HOST_NAME
-  S3_PUBLISH S3_REGION
-  WEB_BROWSER INSTANCE_REGION
-  SHUTDOWN_WAIT PROFILER
-  SIMULATIONS_SPREADSHEET_UPDATE_URL SLACK_HOOK_WITH_TOKEN
-  SIMULATION_HOST_LOG_FILE
-  SIGOPT_CLIENT_ID SIGOPT_DEV_ID  # TODO maybe completely remove
-)
-echo "Following variables might be set only outside of the image (variable name -> 'current value'):"
+required_variables_from_outside=( BEAM_CONFIG MAX_RAM )
+echo "Following variables are required for BEAM to work (variable name -> 'current value'):"
 for v in "${required_variables_from_outside[@]}" ; do
   vval1="$v"
   vval="${!vval1}"
   echo "$v -> '$vval'"
+
+  if [ -z "$vval" ]; then
+    print_error "'$v' environment variable is required for BEAM to work, please, specify a value for it."
+    exit 1
+  fi
 done
-
-
-##
-## fix max ram if not set
-##
-if [[ -z "$MAX_RAM" ]]; then MAX_RAM="8"; echo "MAX_RAM was not set, using default value: '$MAX_RAM'"; fi
 
 
 ##
 ## print optional variables that might be set only outside of image
 ##
-optional_variables_from_outside=( PULL_CODE PULL_DATA SEND_NOTIFICATION )
-echo "Following OPTIONAL variables might be set only outside of the image (variable name -> 'current value'):"
+optional_variables_from_outside=(
+  BEAM_COMMIT_SHA BEAM_DATA_COMMIT_SHA
+  BEAM_BRANCH_NAME BEAM_DATA_BRANCH_NAME
+  PULL_CODE PULL_DATA
+  ENFORCE_HTTPS_FOR_DATA_REPOSITORY
+  S3_PUBLISH AWS_SECRET_ACCESS_KEY AWS_ACCESS_KEY_ID S3_REGION
+  PROFILER
+  SIMULATION_HOST_LOG_FILE
+  GRADLE_CACHE_PATH
+  NOTIFICATION_SEND
+  SIMULATIONS_SPREADSHEET_UPDATE_URL SLACK_HOOK_WITH_TOKEN
+  NOTIFICATION_TITLED NOTIFICATION_SHUTDOWN_WAIT
+  NOTIFICATION_INSTANCE_ID NOTIFICATION_INSTANCE_TYPE NOTIFICATION_HOST_NAME
+  NOTIFICATION_WEB_BROWSER NOTIFICATION_INSTANCE_REGION
+  NOTIFICATION_SIGOPT_CLIENT_ID NOTIFICATION_SIGOPT_DEV_ID
+)
+echo "Following variables are optional (variable name -> 'current value'):"
 for v in "${optional_variables_from_outside[@]}" ; do
   vval1="$v"
   vval="${!vval1}"
-  if [[ -z "$vval" ]]; then
-    eval "$v=true"
-    vval1="$v"
-    vval="${!vval1}"
-    echo "$v -> '$vval'  (using default value)"
-  else
-    echo "$v -> '$vval'"
-  fi
+  echo "$v -> '$vval'"
 done
 
 
@@ -78,49 +83,55 @@ function send_json_to_spreadsheet() {
 
 
 ##
-## logging CPU | RAM usage during simulation
-##
-CPU_RAM_LOG="/app/sources/cpu_ram_usage.csv"
-/app/write_cpu_ram_usage.sh > "$CPU_RAM_LOG" &
-
-
-##
 ## location to project folder
 ##
-PATH_TO_PROJECT_PARENT="/app/sources"
-cd "$PATH_TO_PROJECT_PARENT" || echo "ERROR: The path '$PATH_TO_PROJECT_PARENT' is not available"
+path_to_project_parent="/app/sources"
+cd "$path_to_project_parent" || echo "ERROR: The path '$path_to_project_parent' is not available"
 
 
 ##
 ## either pull the code or use the code from mounted folder
 ##
 if [ "$PULL_CODE" = true ]; then
-  BEAM_NAME="beam"
-  if [ -z "$BEAM_CODE_URL" ]; then BEAM_CODE_URL="https://github.com/LBNL-UCB-STI/beam.git"; fi
-  echo "Pulling the code from github (PULL_CODE set to '$PULL_CODE'), cloning '$BEAM_CODE_URL' into $(pwd)/$BEAM_NAME"
-  git clone --single-branch --branch "$BEAM_BRANCH_NAME" "$BEAM_CODE_URL" "$BEAM_NAME"
-  cd "$BEAM_NAME" || echo "ERROR: The dir '$BEAM_NAME' is not available"
+  if [ -z "$BEAM_BRANCH_NAME" ]; then
+    print_error "PULL_CODE set to true, but BEAM_BRANCH_NAME is empty! Please, specify both."
+    exit 1
+  fi
+
+  #  beam_name="beam"
+  beam_code_url="https://github.com/LBNL-UCB-STI/beam.git"
+  echo "Pulling the code from github (PULL_CODE set to '$PULL_CODE'), cloning '$beam_code_url' into mounted folder"
+  git clone --single-branch --branch "$BEAM_BRANCH_NAME" "$beam_code_url" . || exit 1
+  #  cd "$beam_name" || echo "ERROR: The dir '$beam_name' is not available"
 
   if [[ $BEAM_COMMIT_SHA ]]; then
-    echo "Resetting the branch '$BEAM_BRANCH_NAME' to commit '$BEAM_COMMIT_SHA'"
+    echo "Resetting the code to commit '$BEAM_COMMIT_SHA'"
     git reset --hard "$BEAM_COMMIT_SHA";
   else
     BEAM_COMMIT_SHA=$(git log -1 --pretty=format:%H)
     echo "Using the latest commit in the branch '$BEAM_BRANCH_NAME' - '$BEAM_COMMIT_SHA'"
   fi
 else
-  echo "Pulling the code is disabled (PULL_CODE set to '$PULL_CODE'), using code mounted to '$PATH_TO_PROJECT_PARENT'."
+  echo "Pulling the code is disabled (PULL_CODE set to '$PULL_CODE'), using code mounted to '$path_to_project_parent'."
   BEAM_COMMIT_SHA=$(git log -1 --pretty=format:%H)
   BEAM_BRANCH_NAME=$(git symbolic-ref --short HEAD)
-  echo "Using the latest commit in the branch '$BEAM_BRANCH_NAME' - '$BEAM_COMMIT_SHA'"
+  echo "Using the branch '$BEAM_BRANCH_NAME', commit '$BEAM_COMMIT_SHA'"
 fi
 
 
 ##
-## remember the BEAM working folder location
+## logging CPU | RAM usage during simulation
 ##
-BEAM_PATH=$(pwd)
-echo "Working from '$BEAM_PATH'"
+cpu_ram_log="/app/sources/cpu_ram_usage.csv"
+/app/write_cpu_ram_usage.sh > "$cpu_ram_log" &
+
+
+##
+## Remember the BEAM working folder location.
+## This is required to be able either pull the code or use existing code.
+##
+beam_path=$(pwd)
+echo "Working from '$beam_path'"
 
 
 ##
@@ -136,6 +147,11 @@ fi
 ## pulling data from github if enabled or checking if data location was mounted separately
 ##
 if [ "$PULL_DATA" = true ]; then
+  if [ -z "$BEAM_DATA_BRANCH_NAME" ]; then
+    print_error "PULL_DATA set to true, but BEAM_DATA_BRANCH_NAME is empty! Please, specify both."
+    exit 1
+  fi
+
   echo "Pulling the data from github (PULL_DATA set to '$PULL_DATA')."
   production_data_submodules=$(git submodule | awk '{ print $2 }')
   for i in $production_data_submodules;
@@ -161,13 +177,18 @@ if [ "$PULL_DATA" = true ]; then
   git lfs pull
 else
   echo "Pulling the data from github is disabled (PULL_DATA set to '$PULL_DATA')."
-  COMBINED_CONFIG_PATH="/app/data/${BEAM_CONFIG#*/}"
-  echo "Trying combined path from data volume: '$COMBINED_CONFIG_PATH'"
-  if [ -e "$COMBINED_CONFIG_PATH" ]; then
-    BEAM_CONFIG=$COMBINED_CONFIG_PATH
-    echo "File exist, using config from mounted data volume - '$BEAM_CONFIG'"
+  combined_config_path1="/app/data/${BEAM_CONFIG#*/}"
+  combined_config_path2="/app/sources/${BEAM_CONFIG}"
+  echo "Trying combined path1 from data volume: '$combined_config_path1'"
+  echo "Trying combined path2 from source volume: '$combined_config_path2'"
+  if [ -e "$combined_config_path1" ]; then
+    BEAM_CONFIG=$combined_config_path1
+    echo "File by path1 exist, using absolute path to config from mounted data volume - '$BEAM_CONFIG'"
+  elif [ -e "$combined_config_path2" ]; then
+    echo "File by path2 exist, using relative path to config from source folder - '$BEAM_CONFIG'"
   else
-    echo "File does not exist, going to use data from inside beam folder."
+    print_error "Unable to locate config by path '$BEAM_CONFIG', please specify correct path."
+    exit 1
   fi
 fi
 
@@ -175,36 +196,36 @@ fi
 ##
 ## notification
 ##
-if [ "$SEND_NOTIFICATION" = true ]; then
+if [ "$NOTIFICATION_SEND" = true ]; then
   send_slack_notification "Run Started
-    Run Name $TITLED
-    Instance ID $INSTANCE_ID
-    Instance type $INSTANCE_TYPE
-    Host name $HOST_NAME
-    Web browser $WEB_BROWSER
+    Run Name $NOTIFICATION_TITLED
+    Instance ID $NOTIFICATION_INSTANCE_ID
+    Instance type $NOTIFICATION_INSTANCE_TYPE
+    Host name $NOTIFICATION_HOST_NAME
+    Web browser $NOTIFICATION_WEB_BROWSER
     Branch $BEAM_BRANCH_NAME
     Commit $BEAM_COMMIT_SHA"
 
   send_json_to_spreadsheet "\"status\":\"Run Started\",
-    \"name\":\"$TITLED\",
-    \"instance_id\":\"$INSTANCE_ID\",
-    \"instance_type\":\"$INSTANCE_TYPE\",
-    \"host_name\":\"$HOST_NAME\",
-    \"browser\":\"$WEB_BROWSER\",
+    \"name\":\"$NOTIFICATION_TITLED\",
+    \"instance_id\":\"$NOTIFICATION_INSTANCE_ID\",
+    \"instance_type\":\"$NOTIFICATION_INSTANCE_TYPE\",
+    \"host_name\":\"$NOTIFICATION_HOST_NAME\",
+    \"browser\":\"$NOTIFICATION_WEB_BROWSER\",
     \"branch\":\"$BEAM_BRANCH_NAME\",
     \"commit\":\"$BEAM_COMMIT_SHA\",
     \"data_branch\":\"$BEAM_DATA_BRANCH_NAME\",
     \"data_commit\":\"$BEAM_DATA_COMMIT_SHA\",
-    \"region\":\"$INSTANCE_REGION\",
+    \"region\":\"$NOTIFICATION_INSTANCE_REGION\",
     \"batch\":\"\",
     \"s3_link\":\"\",
     \"max_ram\":\"$MAX_RAM\",
     \"profiler_type\":\"$PROFILER\",
     \"config_file\":\"$BEAM_CONFIG\",
-    \"sigopt_client_id\":\"$SIGOPT_CLIENT_ID\",
-    \"sigopt_dev_id\":\"$SIGOPT_DEV_ID\""
+    \"sigopt_client_id\":\"$NOTIFICATION_SIGOPT_CLIENT_ID\",
+    \"sigopt_dev_id\":\"$NOTIFICATION_SIGOPT_DEV_ID\""
 else
-  echo "Sending notifications is disabled (SEND_NOTIFICATION set to '$SEND_NOTIFICATION')."
+  echo "Sending notifications is disabled (NOTIFICATION_SEND set to '$NOTIFICATION_SEND')."
 fi
 
 
@@ -212,8 +233,8 @@ fi
 ## calculating a location for gradle cache
 ##
 if [[ -z "$GRADLE_CACHE_PATH" ]]; then
-  GRADLE_CACHE_PATH="$BEAM_PATH/.gradle"
-  echo "GRADLE_CACHE_PATH is not set, creating and using directory by path '$GRADLE_CACHE_PATH'"
+  GRADLE_CACHE_PATH="$beam_path/.gradle"
+  echo "GRADLE_CACHE_PATH is not set, using directory by path '$GRADLE_CACHE_PATH'"
   mkdir -p "$GRADLE_CACHE_PATH"
 else
   echo "GRADLE_CACHE_PATH set, using directory by path '$GRADLE_CACHE_PATH'"
@@ -230,7 +251,7 @@ fi
 ##
 ## Calculate the final status of simulation
 ##
-log_file="$(find "$BEAM_PATH/output" -maxdepth 2 -mindepth 2 -type d -print -quit)/beamLog.out"
+log_file="$(find "$beam_path/output" -maxdepth 2 -mindepth 2 -type d -print -quit)/beamLog.out"
 if [[ ! -f $log_file ]]; then
     echo "Unable to locate the beamLog.out file"
     final_status="Unable to start"
@@ -256,7 +277,6 @@ health_metrics=""
 simulation_health_analysis_output_file="simulation_health_analysis_result.txt"
 python3 src/main/python/general_analysis/simulation_health_analysis.py $simulation_health_analysis_output_file
 while IFS="," read -r metric count; do
-  echo "Health metric exporting: '$metric=$count'"
   export "$metric=$count"
   health_metrics="$health_metrics, $metric:$count"
 done < $simulation_health_analysis_output_file
@@ -277,24 +297,28 @@ done;
 echo "Found output dir: $FINAL_PATH"
 
 echo "Moving debug files to output folder."
-for file in "$BEAM_PATH"/*.jfr; do
+for file in "$beam_path"/*.jfr; do
   [ -e "$file" ] || continue
   echo "Zipping $file"
   zip "$file.zip" "$file"
   mv "$file.zip" "$FINAL_PATH"
 done;
 
-echo "Copy health metrics."
-cp "$simulation_health_analysis_output_file" "$FINAL_PATH"
+echo "Moving health metrics."
+mv "$simulation_health_analysis_output_file" "$FINAL_PATH"
 
-echo "Copy garbage collection logs."
-cp "$BEAM_PATH"/gc_* "$FINAL_PATH"
+echo "Moving garbage collection logs."
+mv "$beam_path"/gc_* "$FINAL_PATH"
 
-echo "Copy CPU and RAM used/available collected during simulation."
-gzip "$CPU_RAM_LOG" && cp "$CPU_RAM_LOG"* "$FINAL_PATH"
+echo "Moving CPU and RAM used/available collected during simulation."
+gzip "$cpu_ram_log" && mv "$cpu_ram_log"* "$FINAL_PATH"
 
-echo "Copy log of simulation from host computer (cloud-init log or analog)."
-cp "$SIMULATION_HOST_LOG_FILE" "$FINAL_PATH"
+if [ -n "$SIMULATION_HOST_LOG_FILE" ]; then
+  echo "Copy log of simulation from host computer (cloud-init log or analog)."
+  cp "$SIMULATION_HOST_LOG_FILE" "$FINAL_PATH"
+else
+  echo "Path to host computer simulation log file (cloud-init log or analog) was not provided."
+fi
 
 echo "Fixing permission issues related to access to files created from inside of image."
 chmod 777 -R "$FINAL_PATH"
@@ -306,7 +330,7 @@ chmod 777 -R "$FINAL_PATH"
 if [ "$S3_PUBLISH" = true ]; then
   aws --region "$S3_REGION" s3 cp "$FINAL_PATH" s3://beam-outputs/"$FINAL_PATH" --recursive;
   s3output_url="https://s3.$S3_REGION.amazonaws.com/beam-outputs/index.html#$FINAL_PATH"
-  SIMULATION_OUTPUT_LINK="$s3output_url"
+  simulation_output_link="$s3output_url"
   echo "Uploaded to $s3output_url"
 else
   echo "S3 publishing is disabled (S3_PUBLISH set to '$S3_PUBLISH')."
@@ -316,33 +340,33 @@ fi
 ##
 ## notification
 ##
-if [ "$SEND_NOTIFICATION" = true ]; then
+if [ "$NOTIFICATION_SEND" = true ]; then
   send_slack_notification "Run Completed
-    Run Name $TITLED
-    Instance ID $INSTANCE_ID
-    Instance type $INSTANCE_TYPE
-    Host name $HOST_NAME
-    Web browser $WEB_BROWSER
+    Run Name $NOTIFICATION_TITLED
+    Instance ID $NOTIFICATION_INSTANCE_ID
+    Instance type $NOTIFICATION_INSTANCE_TYPE
+    Host name $NOTIFICATION_HOST_NAME
+    Web browser $NOTIFICATION_WEB_BROWSER
     Branch $BEAM_BRANCH_NAME
     Commit $BEAM_COMMIT_SHA
     Health Metrics $health_metrics
     S3 output url $s3output_url
-    Shutdown in $SHUTDOWN_WAIT minutes"
+    Shutdown in $NOTIFICATION_SHUTDOWN_WAIT minutes"
 
   # shellcheck disable=SC2089
   send_json_to_spreadsheet "\"status\":\"$final_status\",
-    \"name\":\"$TITLED\",
-    \"instance_id\":\"$INSTANCE_ID\",
-    \"instance_type\":\"$INSTANCE_TYPE\",
-    \"host_name\":\"$HOST_NAME\",
-    \"browser\":\"$WEB_BROWSER\",
+    \"name\":\"$NOTIFICATION_TITLED\",
+    \"instance_id\":\"$NOTIFICATION_INSTANCE_ID\",
+    \"instance_type\":\"$NOTIFICATION_INSTANCE_TYPE\",
+    \"host_name\":\"$NOTIFICATION_HOST_NAME\",
+    \"browser\":\"$NOTIFICATION_WEB_BROWSER\",
     \"branch\":\"$BEAM_BRANCH_NAME\",
     \"commit\":\"$BEAM_COMMIT_SHA\",
     \"data_branch\":\"$BEAM_DATA_BRANCH_NAME\",
     \"data_commit\":\"$BEAM_DATA_COMMIT_SHA\",
-    \"region\":\"$INSTANCE_REGION\",
+    \"region\":\"$NOTIFICATION_INSTANCE_REGION\",
     \"batch\":\"\",
-    \"s3_link\":\"$SIMULATION_OUTPUT_LINK\",
+    \"s3_link\":\"$simulation_output_link\",
     \"max_ram\":\"$MAX_RAM\",
     \"profiler_type\":\"$PROFILER\",
     \"stacktrace\":\"$stacktrace\",
@@ -350,10 +374,10 @@ if [ "$SEND_NOTIFICATION" = true ]; then
     \"error\":\"$error\",
     \"warning\":\"$warn\",
     \"config_file\":\"$BEAM_CONFIG\",
-    \"sigopt_client_id\":\"$SIGOPT_CLIENT_ID\",
-    \"sigopt_dev_id\":\"$SIGOPT_DEV_ID\""
+    \"sigopt_client_id\":\"$NOTIFICATION_SIGOPT_CLIENT_ID\",
+    \"sigopt_dev_id\":\"$NOTIFICATION_SIGOPT_DEV_ID\""
 else
-  echo "Sending notifications is disabled (SEND_NOTIFICATION set to '$SEND_NOTIFICATION')."
+  echo "Sending notifications is disabled (NOTIFICATION_SEND set to '$NOTIFICATION_SEND')."
 fi
